Move LineFunc class to separate file [Images/ImageStuff/PositionManager.cs, MathStuff/LineFunc.cs]
